+++ logprep/abc/processor.py
@@ -205,6 +205,8 @@ class Processor(Component):
             self._apply_rules(event, rule)
         except ProcessingWarning as error:
             self._handle_warning_error(event, rule, error)
+        except ProcessingCriticalError as error:
+            raise error
         except BaseException as error:
             raise ProcessingCriticalError(self, str(error), event) from error
         if not hasattr(rule, "delete_source_fields"):