+++ tests/acceptance/test_amides.py
@@ -2,13 +2,12 @@
 # pylint: disable=not-an-iterable
 # pylint: disable=missing-docstring
 
-from logging import getLogger, DEBUG, basicConfig
+from logging import DEBUG, basicConfig, getLogger
 
 import pytest
 
 from logprep.util.json_handling import dump_config_as_file
 from tests.acceptance.util import get_test_output
-
 
 basicConfig(level=DEBUG, format="%(asctime)-15s %(name)-5s %(levelname)-8s: %(message)s")
 logger = getLogger("Logprep-Test")
@@ -55,7 +54,9 @@ def test_amides(tmp_path, config):
     dump_config_as_file(config_path, config)
 
     test_output = get_test_output(config_path)
-    test_output_documents = [
-        event for event in test_output[0] if event.get("rule_attributions", None)
+    test_output_documents = [event for event in test_output[0] if event.get("amides")]
+    attributed_documents = [
+        event for event in test_output_documents if event.get("amides").get("attributions")
     ]
-    assert len(test_output_documents) == 8
+    assert len(test_output_documents) == 20
+    assert len(attributed_documents) == 8