+++ tests/unit/processor/dissector/test_dissector_rule.py
@@ -74,8 +74,8 @@ class TestDissectorRule:
                     "filter": "message",
                     "dissector": {"mapping": {"field": "%{ts}"}},
                 },
-                ValueError,
-                "must match regex",
+                None,
+                None,
             ),
             (
                 {